Flag smoke puffs left behind by enemies as "just moved"
This resolves several problems with update order, that allowed players
to end up on the same tile as an enemy, getting damaged every beat
until dead. This may somewhat alter behavior of enemies that move near
one another, but it should on the whole result in more consistent
resolution of those collisions, rather than the old behavior which
heavily favored leftward / upwards movement.