Port the Action53 codebase to Rainbow with very minimal changes
Here I'm having Rainbow replicate the memory layout of the Action53
cart, which allows the code to mostly remain untouched. Eventually
I want to switch to a more competent memory layout (8k PRG banks)
and start using the extended CHR banking features of the FPGA RAM,
but for now this will do as a starting point.