Rework all debugging features to be invisible, even when toggled on
The Lua script in Mesen can perfectly well respond to writes targeting
registers that don't exist. We'll just do that, and not blind ourselves
with emphasis+greyscale changes during gameplay.

Also the debug bits of the profiling script can now be toggled on and
off, which helps with setting up the scenario that you want to test
while still being able to see what you're doing. Useful, that!