Spit out a list of definitions for the engine to consume
Also this marks a v0.96 for the CHR ROM generation: all assets
are present in theory, and now it's a matter of reworking the
engine to actually consume them. That's more work than it sounds
like due to hard-coded tile IDs everywhere, but we only need to
fix that once in theory.