Add a debug cookie and a set of "blocking" test chambers
These are intentionally a little bit ugly and VERY simple. We'll
use these when planning the game's high level progression, so we
don't have to block ourselves on artwork while we are working on
the basic mechanics.

Better to figure out that we want to cut an area for game feel
BEFORE we spend mountains of time on the tiles.