Implement semisafe targeting for the new system
It's still possible to kite enemies into getting stuck on semisafe
tiles, if that's the "closest" path towards the player, but it
isn't quite as cheesy as the old system. Fixing it properly is
extremely tricky due to not knowing the player's direction before
the pathfinding is done, and semisafe tiles are rare enough that
I frankly don't mind a bit of exploitation on the player's part.
None of the really important combat encounters will feature them.