Mushroom enemies now clean up their spores if defeated
The guiding principle here is that a player can "interrupt" and
"cancel" an attack by defeating the enemy on the turn "before" the
attack happens. For several types of attacks, this is more
complicated than it appears, but we do want to be very consistent
with the behavior. Always judge in the player's favor.