Add a "patient" mode, letting the player move at their own pace
This is intended as an accessibility mode, and we will not default
to it. Helpful for the hard of hearing or rhythmically challenged.
Right now it does not permit "waiting" on any given beat, giving
the mode some additional nuance. We probably want to fix this at
some point so that it has strategic parity with the other game
modes.

Later we could use this for a puzzle-type mode!