Wait to update the animation frame until the end of the game loop
Processing it so early was causing actual lag to usually interrupt
the sprite/torchlight routine, and this rendered the new nametable
with the old animation frame, which was particularly disasterous
on the first frame of the beat. Unfortunately the player's frame
is at high risk to lag with all the extra logic, and so we need
to prioritize this specific frame looking good. This doesn't
eliminate the risk of a desync, but it does lower the chances
considerably.