color -> classical element
I'm a bit torn on water/ice in particular, but we'll just have
to run with that I guess. Air is a weaker thunder. I dunno.
Anyway, the other thing is to make it more obvious when I'm
trying to use a specific palette index or a specific element,
since many parts of the HUD and UI are using completely different
palettes.