Organize enemy functions by TILE_ID rather than by dispatch
This puts all related functions for a single enemy close together,
which is muuuuch easier to manage, and was rather the point of all
this organizational fussing. Now adding new enemy types is less
stupid in general, and there are guardrails to stop me from screwing
up the order and not at least realizing what I've done.