Massage save file format with reserved / padding bytes
We decided early on that 128 bytes ought to be plenty, and we're
pretty sure we are keeping all of the currently defined bytes.
This gives us some room to expand without changing the underlying
struct size or shape, which should allow us to add options without
breaking old saves files in the process. Very handy for testing.