Hookup the basics of colorspace lookup tables for rooms.
With this, technically protan/tritan modes are functional, though
only the grasslands room has a proper palette entry in the set.
Now we mostly just need to run through and flesh out the rest of
the variant palettes. There are a bunch more edge cases near the
other parts of the UI, most notably temporary hearts in the HUD
in the two monochrome modes.