Begin work on localization. Localized signs work! UI needs work.
Supported languages will be two forms of Toki Pona for a good long
while, as I don't want to get other translators involved until the
script is more or less finalized. It'll still take a ton of effort
to rework all string drawing into localized forms, especially for
the UI, so it's good to get that out of the way early.