Take a first stab at hooking up bandit rewards
There's a LOT of jank, and I need to stop and go to bed. The basics
are mostly working, spellcasting is broken, and drawing tasks are
especially cursed when spawning new items and whatnot. The core of
it is doing what I want, it just needs a lot of polish before I'd
be comfortable showing it off.