Defer tile updating until the following beat
It turns out death sprites were deferred and I don't quite remember
how, but it looks better this way. Items also spawn on visual defeat
rather than a whole-ass beat later, which is muuuuch less broken
looking. The main thing we still want here is some sort of visual
confirmation of the player's successful activation? Maybe a special
chime SFX (rather than defeat) or something to that effect. Can we
defer a SFX to the next beat? That would be a nifty mechanic.