Hookup cultist spellcast behavior
This is similar to most regular enemies, except that we want
cultists to interrupt any ongoing spellcast, as the player doesn't
have enough time to react to it while they've committed to their
own spell. This also dodges a bug: we won't have a cultist prepare
a spell in their original color, then cast it in their new color
with mismatched warning tiles.

Opportunities for exploitation here are rare, and mostly the jank
resolves in the player's favor, which we accept.