Nerf the Defensive Shield (to make room for new defense items)
This was too powerful as an all-around damage reduction. We want
to heavily encourage the player to specialize, to spice up build
strategies, so they should always be more able to defend against
SPECIFIC sources of damage than the general all-rounder items
provide. We may lower the price to compensate, although paired
with heart armor this is still a *solid* jack-of-all-trades choice
for the mid game. (It definitely makes heart armor much more
valueable again.)