Have charge resolve on the beat, without waiting for an input
This fixes a minor polish feel if the player releases the button
to charge, which is a natural thing to try to do if you've played
other games. We don't require this but we do care if it screws
with the apparent beat timing. When you charge in this game you
**commit** to it, and that committment can sync to the music.