Greatly rebalance earlygame loot for mostly demo purposes
Trying to heavily prioritize getting the player some interesting
mechanics to play with right now, as we are in demo mode and need
to set hooks. Later, especially once we have more than a single
zone fleshed out, we'll need to redo this and slow the growth
down a bit. (In general, the first zone is too large and too
full of tanky foes; it slows down the pace.)